+++ .babelrc
@@ -1,6 +1,7 @@
 {
   "presets": ["@babel/preset-env", "@babel/preset-react"],
   "plugins": [
+    "@babel/transform-runtime",
     "@babel/plugin-proposal-class-properties",
     [
       "@babel/plugin-proposal-decorators",
