+++ package.json
@@ -16,6 +16,7 @@
     "coverage": "jest --coverage"
   },
   "dependencies": {
+    "@babel/runtime": "^7.10.2",
     "@material-ui/core": "^4.10.0",
     "@material-ui/icons": "^4.9.1",
     "@openemp-mf/styleguide": "github:honohunter/openemp-ui-styleguide#alpha",
@@ -34,6 +35,7 @@
     "@babel/plugin-proposal-class-properties": "^7.8.3",
     "@babel/plugin-proposal-decorators": "^7.8.3",
     "@babel/plugin-proposal-object-rest-spread": "7.9.6",
+    "@babel/plugin-transform-runtime": "^7.10.1",
     "@babel/preset-env": "^7.9.6",
     "@babel/preset-react": "^7.9.4",
     "@testing-library/react": "^10.0.4",
