+++ src/components/desktopSection.jsx
@@ -18,12 +18,12 @@ export default function DesktopSection({ menuId, handleProfileMenuOpen }) {
     <div className={classes.sectionDesktop}>
       <IconButton aria-label="show 4 new mails" color="inherit">
         <Badge badgeContent={4} color="secondary">
-          <Icon class="material-icons">mail</Icon>
+          <Icon>mail</Icon>
         </Badge>
       </IconButton>
       <IconButton aria-label="show 17 new notifications" color="inherit">
         <Badge badgeContent={17} color="secondary">
-          <Icon class="material-icons">notifications</Icon>
+          <Icon>notifications</Icon>
         </Badge>
       </IconButton>
       <IconButton
@@ -34,7 +34,7 @@ export default function DesktopSection({ menuId, handleProfileMenuOpen }) {
         onClick={handleProfileMenuOpen}
         color="inherit"
       >
-        <Icon class="material-icons">account_circle</Icon>
+        <Icon>account_circle</Icon>
       </IconButton>
     </div>
   );
