+++ src/components/index.js
@@ -4,3 +4,4 @@ export { default as ProfileMenu } from './profileMenu';
 export { default as Search } from './search';
 export { default as DesktopSection } from './desktopSection';
 export { default as MobileSection } from './mobileSection';
+export { default as NavbarFrame } from './navbarFrame';
