+++ src/features/index.js
@@ -1,2 +1,2 @@
 /* eslint-disable import/prefer-default-export */
-export { default as Navbar } from './navbar';
+export { default as DrawerControl } from './drawerControl';
